Update README.md
My personal website is created using HTML, JS, and CSS. It showcases my portfolio and provides information about my skills and experiences. The design is clean and minimalist, making it easy for visitors to navigate. I regularly update the site with new projects and blog posts related to my interests in web development and technology. The site is optimized for both desktop and mobile devices, ensuring that visitors can access it from anywhere. With my personal website, I aim to showcase my abilities as a web developer and connect with like-minded individuals in the tech industry.